+++ src/main/java/project/realtimechatapplication/repository/jpa/NotificationRepository.java
@@ -1,4 +1,4 @@
-package project.realtimechatapplication.repository;
+package project.realtimechatapplication.repository.jpa;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import project.realtimechatapplication.entity.NotificationEntity;
