+++ src/main/java/project/realtimechatapplication/repository/jpa/ReactionRepository.java
@@ -1,4 +1,4 @@
-package project.realtimechatapplication.repository;
+package project.realtimechatapplication.repository.jpa;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
