+++ src/test/java/project/realtimechatapplication/service/AuthServiceImplTest.java
@@ -36,8 +36,8 @@ import project.realtimechatapplication.exception.impl.VerificationNumberNotMatch
 import project.realtimechatapplication.exception.impl.WrongPasswordException;
 import project.realtimechatapplication.provider.EmailProvider;
 import project.realtimechatapplication.provider.TokenProvider;
-import project.realtimechatapplication.repository.UserRepository;
-import project.realtimechatapplication.repository.VerificationRepository;
+import project.realtimechatapplication.repository.jpa.UserRepository;
+import project.realtimechatapplication.repository.jpa.VerificationRepository;
 import project.realtimechatapplication.service.impl.AuthServiceImpl;
 
 public class AuthServiceImplTest {
